Refactor: Move auth and transaction logic to controllers and update routes to use them
