+++ backend/src/controllers/transactionController.js
@@ -1,0 +1,69 @@
+// Controller for transaction logic
+const { transactions } = require('../shared/data');
+
+// Add income transaction
+exports.addIncome = (req, res) => {
+  const { amount, category, description, date } = req.body;
+  if (!amount || amount <= 0) {
+    return res.status(400).json({ success: false, error: 'Amount must be positive' });
+  }
+  const transaction = {
+    id: Date.now().toString(),
+    userId: req.userId,
+    type: 'income',
+    amount,
+    category,
+    description,
+    date: date || new Date(),
+    createdAt: new Date()
+  };
+  transactions.push(transaction);
+  res.status(201).json({ success: true, message: 'Income transaction added successfully', data: transaction });
+};
+
+// Add expense transaction
+exports.addExpense = (req, res) => {
+  const { amount, category, description, date } = req.body;
+  if (!amount || amount <= 0) {
+    return res.status(400).json({ success: false, error: 'Amount must be greater than 0' });
+  }
+  if (!category) {
+    return res.status(400).json({ success: false, error: 'Category is required' });
+  }
+  const transaction = {
+    id: Date.now().toString(),
+    userId: req.userId,
+    type: 'expense',
+    amount,
+    category,
+    description,
+    date: date || new Date(),
+    createdAt: new Date()
+  };
+  transactions.push(transaction);
+  res.status(201).json({ success: true, message: 'Expense transaction added successfully', data: transaction });
+};
+
+// Get income summary
+exports.getIncomeSummary = (req, res) => {
+  const userIncome = transactions.filter(t => t.userId === req.userId && t.type === 'income');
+  const totalIncome = userIncome.reduce((sum, t) => sum + t.amount, 0);
+  res.status(200).json({ success: true, data: { totalIncome, incomeCount: userIncome.length, transactions: userIncome } });
+};
+
+// Get expense summary
+exports.getExpenseSummary = (req, res) => {
+  const userExpenses = transactions.filter(t => t.userId === req.userId && t.type === 'expense');
+  const totalExpenses = userExpenses.reduce((sum, t) => sum + t.amount, 0);
+  res.status(200).json({ success: true, data: { totalExpenses, expenseCount: userExpenses.length, transactions: userExpenses } });
+};
+
+// Get expenses (optionally filtered by category)
+exports.getExpenses = (req, res) => {
+  const { category } = req.query;
+  let userExpenses = transactions.filter(t => t.userId === req.userId && t.type === 'expense');
+  if (category) {
+    userExpenses = userExpenses.filter(t => t.category === category);
+  }
+  res.status(200).json({ success: true, data: { transactions: userExpenses } });
+};
